Add missing cast to BYTE

+++ src/Native/Runtime/windows/PalRedhawkMinWin.cpp
@@ -1373,7 +1373,7 @@ bool GCToOSInterface::SetCurrentThreadIdealAffinity(GCThreadAffinity* affinity)
     {
         if (GetThreadIdealProcessorEx(GetCurrentThread(), &proc))
         {
-            proc.Number = affinity->Processor;
+            proc.Number = (BYTE)affinity->Processor;
             success = !!SetThreadIdealProcessorEx(GetCurrentThread(), &proc, NULL);
         }        
     }
